refactor: remove circular dependency in payment-input-validator
