+++ src/domain/wallets/payment-input-validator.ts
@@ -1,8 +1,10 @@
 import { AccountStatus } from "@domain/accounts"
 import { InactiveAccountError, InvalidWalletId, SelfPaymentError } from "@domain/errors"
 import { checkedToBtcPaymentAmount, checkedToUsdPaymentAmount } from "@domain/payments"
+
 import { WalletCurrency } from "@domain/shared"
-import { checkedToWalletId } from "@domain/wallets"
+
+import { checkedToWalletId } from "./validation"
 
 export const PaymentInputValidator = (
   getWalletFn: PaymentInputValidatorConfig,
