+++ src/domain/wallets/validation.ts
@@ -1,0 +1,13 @@
+import { InvalidWalletId } from "@domain/errors"
+
+const UuidRegex =
+  /^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/i
+
+export const WalletIdRegex = UuidRegex
+
+export const checkedToWalletId = (walletId: string): WalletId | ValidationError => {
+  if (!walletId.match(WalletIdRegex)) {
+    return new InvalidWalletId(walletId)
+  }
+  return walletId as WalletId
+}
